requirements.txt
streamlit
google-generativeai
gtts
streamlit-webrtc
openai-whisper
soundfile